refactor: split nature.ts into 6 domain files + barrel (#476)
Moves fruits, vegetables, animals (+ MEMORY_GAME_ANIMALS + ANIMAL_SOUND_FREQUENCIES),
smellTaste, oceanLife, gardenPlants into natureData/ subfiles. nature.ts
becomes a 6-line barrel — zero breaking changes for consumers.
