+++ gamesformykids/lib/constants/gameData/natureData/animals.ts
@@ -1,0 +1,62 @@
+import { BaseGameItem } from "@/lib/types/core/base";
+import { createItemsList, createPronunciationDictionary, DEFAULT_GAME_CONFIG } from "@/lib/constants/core";
+
+export const ANIMALS: Record<string, BaseGameItem> = {
+  DOG: { name: "dog", hebrew: "כלב", english: "Dog", emoji: "🐶", color: "bg-brown-500", sound: [200, 300, 150] },
+  CAT: { name: "cat", hebrew: "חתול", english: "Cat", emoji: "🐱", color: "bg-gray-500", sound: [800, 1000, 600] },
+  COW: { name: "cow", hebrew: "פרה", english: "Cow", emoji: "🐄", color: "bg-pink-300", sound: [100, 200, 150] },
+  HORSE: { name: "horse", hebrew: "סוס", english: "Horse", emoji: "🐴", color: "bg-amber-600", sound: [300, 500, 400] },
+  SHEEP: { name: "sheep", hebrew: "כבש", english: "Sheep", emoji: "🐑", color: "bg-gray-200", sound: [400, 600, 500] },
+  PIG: { name: "pig", hebrew: "חזיר", english: "Pig", emoji: "🐷", color: "bg-pink-400", sound: [250, 350, 200] },
+  CHICKEN: { name: "chicken", hebrew: "תרנגולת", english: "Chicken", emoji: "🐔", color: "bg-yellow-400", sound: [600, 800, 700] },
+  DUCK: { name: "duck", hebrew: "ברווז", english: "Duck", emoji: "🦆", color: "bg-blue-300", sound: [500, 700, 600] },
+  RABBIT: { name: "rabbit", hebrew: "ארנב", english: "Rabbit", emoji: "🐰", color: "bg-gray-300", sound: [400, 500, 600] },
+  FROG: { name: "frog", hebrew: "צפרדע", english: "Frog", emoji: "🐸", color: "bg-green-400", sound: [200, 400, 300] },
+};
+
+export const ALL_ANIMALS = createItemsList(ANIMALS);
+export const ANIMAL_HEBREW_PRONUNCIATIONS = createPronunciationDictionary(ANIMALS);
+export const ANIMAL_GAME_CONSTANTS = DEFAULT_GAME_CONFIG;
+
+export const MEMORY_GAME_ANIMALS = [
+  { emoji: "🐱", sound: "meow", name: "חתול" },
+  { emoji: "🐶", sound: "woof", name: "כלב" },
+  { emoji: "🐰", sound: "hop", name: "ארנב" },
+  { emoji: "🦊", sound: "yip", name: "שועל" },
+  { emoji: "🐻", sound: "growl", name: "דוב" },
+  { emoji: "🐼", sound: "chirp", name: "פנדה" },
+  { emoji: "🐯", sound: "roar", name: "נמר" },
+  { emoji: "🦁", sound: "roar", name: "אריה" },
+  { emoji: "🐸", sound: "ribbit", name: "צפרדע" },
+  { emoji: "🐷", sound: "oink", name: "חזיר" },
+  { emoji: "🐵", sound: "ooh", name: "קוף" },
+  { emoji: "🐨", sound: "growl", name: "קואלה" },
+  { emoji: "🦒", sound: "bleat", name: "ג'ירפה" },
+  { emoji: "🐘", sound: "trumpet", name: "פיל" },
+  { emoji: "🦓", sound: "neigh", name: "זברה" },
+  { emoji: "🦏", sound: "snort", name: "קרנף" },
+  { emoji: "🐄", sound: "moo", name: "פרה" },
+  { emoji: "🐎", sound: "neigh", name: "סוס" },
+];
+
+export const ANIMAL_SOUND_FREQUENCIES: Record<string, number[]> = {
+  "🐱": [800, 1000, 600],
+  "🐶": [200, 300, 150],
+  "🐰": [400, 500, 600],
+  "🦊": [600, 800, 500],
+  "🐻": [100, 150, 80],
+  "🐼": [300, 400, 350],
+  "🐯": [150, 250, 100],
+  "🦁": [120, 200, 90],
+  "🐸": [500, 700, 400],
+  "🐷": [250, 350, 200],
+  "🐵": [450, 550, 350],
+  "🐨": [180, 280, 120],
+  "🦒": [350, 450, 250],
+  "🐘": [80, 120, 60],
+  "🦓": [220, 320, 180],
+  "🦏": [110, 180, 70],
+  "🐄": [160, 240, 100],
+  "🐎": [200, 300, 150],
+  "default": [440, 550, 330],
+};
